+++ app/serializers/cla_user_serializer.rb
@@ -1,5 +1,6 @@
 class ClaUserSerializer < ActiveModel::Serializer
-  attributes :id, :email, :cla_cohort_id, :cla_role_id, :name, :phone_number, :birthday, :course_completion_rate
+  attributes :id, :email, :cla_cohort_id, :cla_role_id, :name, :phone_number, :birthday, :course_completion_rate,
+   :cla_cohort_name, :cla_role_name
 
   def cla_cohort_name
     object.cla_cohort&.name
@@ -7,10 +8,5 @@ class ClaUserSerializer < ActiveModel::Serializer
 
   def cla_role_name
     object.cla_role&.name
-  end
-
-  # Get course completion rate dynamically
-  def course_completion_rate
-    object.course_completion_rate(object.cla_cohort_id)
   end
 end
